changed structure of rest dto packages
