+++ src/main/java/com/knighttodo/knighttodo/rest/dto/todo/request/DeleteTodoRequestDto.java
@@ -1,4 +1,4 @@
-package com.knighttodo.knighttodo.rest.dto.request.todo;
+package com.knighttodo.knighttodo.rest.dto.todo.request;
 
 import lombok.Data;
