+++ src/main/java/com/knighttodo/knighttodo/rest/dto/todo/request/GetTodoRequestDto.java
@@ -1,11 +1,11 @@
-package com.knighttodo.knighttodo.rest.dto.request.todoblock;
+package com.knighttodo.knighttodo.rest.dto.todo.request;
 
 import lombok.Data;
 
 import javax.validation.constraints.Min;
 
 @Data
-public class DeleteTodoBlockRequestDto {
+public class GetTodoRequestDto {
 
     @Min(1)
     private long id;
