+++ src/main/java/com/knighttodo/knighttodo/rest/dto/todo/request/UpdateTodoRequestDto.java
@@ -1,4 +1,4 @@
-package com.knighttodo.knighttodo.rest.dto.response.todo;
+package com.knighttodo.knighttodo.rest.dto.todo.request;
 
 import com.knighttodo.knighttodo.gateway.privatedb.representation.TodoBlock;
 import com.knighttodo.knighttodo.gateway.privatedb.representation.enums.Hardness;
@@ -8,21 +8,30 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.Min;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class UpdateTodoResponseDto {
+public class UpdateTodoRequestDto {
 
+    @Min(1)
     private long id;
 
+    @NotBlank
     private String todoName;
 
+    @NotNull
     private Scaryness scaryness;
 
+    @NotNull
     private Hardness hardness;
 
     private boolean ready;
 
+    @NotNull
     private TodoBlock todoBlock;
 }
