+++ src/main/java/com/knighttodo/knighttodo/rest/dto/todo/response/TodoResponseDto.java
@@ -1,4 +1,4 @@
-package com.knighttodo.knighttodo.rest.dto.response.todo;
+package com.knighttodo.knighttodo.rest.dto.todo.response;
 
 import com.knighttodo.knighttodo.gateway.privatedb.representation.TodoBlock;
 import com.knighttodo.knighttodo.gateway.privatedb.representation.enums.Hardness;
