+++ src/main/java/com/knighttodo/knighttodo/rest/dto/todo/response/UpdateTodoResponseDto.java
@@ -1,4 +1,4 @@
-package com.knighttodo.knighttodo.rest.dto.request.todo;
+package com.knighttodo.knighttodo.rest.dto.todo.response;
 
 import com.knighttodo.knighttodo.gateway.privatedb.representation.TodoBlock;
 import com.knighttodo.knighttodo.gateway.privatedb.representation.enums.Hardness;
@@ -8,30 +8,21 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class UpdateTodoRequestDto {
+public class UpdateTodoResponseDto {
 
-    @Min(1)
     private long id;
 
-    @NotBlank
     private String todoName;
 
-    @NotNull
     private Scaryness scaryness;
 
-    @NotNull
     private Hardness hardness;
 
     private boolean ready;
 
-    @NotNull
     private TodoBlock todoBlock;
 }
