+++ src/main/java/com/knighttodo/knighttodo/rest/dto/todoblock/request/CreateTodoBlockRequestDto.java
@@ -1,4 +1,4 @@
-package com.knighttodo.knighttodo.rest.dto.response.todoblock;
+package com.knighttodo.knighttodo.rest.dto.todoblock.request;
 
 import com.knighttodo.knighttodo.gateway.privatedb.representation.Todo;
 import lombok.AllArgsConstructor;
@@ -6,17 +6,19 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class CreateTodoBlockResponseDto {
+public class CreateTodoBlockRequestDto {
 
-    private long id;
-
+    @NotBlank
     private String blockName;
 
+    @NotNull
     private List<Todo> todos;
 }
