+++ src/main/java/com/knighttodo/knighttodo/rest/dto/todoblock/request/DeleteTodoBlockRequestDto.java
@@ -1,11 +1,11 @@
-package com.knighttodo.knighttodo.rest.dto.request.todo;
+package com.knighttodo.knighttodo.rest.dto.todoblock.request;
 
 import lombok.Data;
 
 import javax.validation.constraints.Min;
 
 @Data
-public class GetTodoRequestDto {
+public class DeleteTodoBlockRequestDto {
 
     @Min(1)
     private long id;
