+++ src/main/java/com/knighttodo/knighttodo/rest/dto/todoblock/response/CreateTodoBlockResponseDto.java
@@ -1,4 +1,4 @@
-package com.knighttodo.knighttodo.rest.dto.request.todoblock;
+package com.knighttodo.knighttodo.rest.dto.todoblock.response;
 
 import com.knighttodo.knighttodo.gateway.privatedb.representation.Todo;
 import lombok.AllArgsConstructor;
@@ -6,19 +6,17 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class CreateTodoBlockRequestDto {
+public class CreateTodoBlockResponseDto {
 
-    @NotBlank
+    private long id;
+
     private String blockName;
 
-    @NotNull
     private List<Todo> todos;
 }
