+++ src/main/java/com/knighttodo/knighttodo/rest/mapper/TodoBlockRestMapper.java
@@ -1,11 +1,11 @@
 package com.knighttodo.knighttodo.rest.mapper;
 
 import com.knighttodo.knighttodo.domain.TodoBlockVO;
-import com.knighttodo.knighttodo.rest.dto.request.todoblock.CreateTodoBlockRequestDto;
-import com.knighttodo.knighttodo.rest.dto.request.todoblock.UpdateTodoBlockRequestDto;
-import com.knighttodo.knighttodo.rest.dto.response.todoblock.CreateTodoBlockResponseDto;
-import com.knighttodo.knighttodo.rest.dto.response.todoblock.TodoBlockResponseDto;
-import com.knighttodo.knighttodo.rest.dto.response.todoblock.UpdateTodoBlockResponseDto;
+import com.knighttodo.knighttodo.rest.dto.todoblock.request.CreateTodoBlockRequestDto;
+import com.knighttodo.knighttodo.rest.dto.todoblock.request.UpdateTodoBlockRequestDto;
+import com.knighttodo.knighttodo.rest.dto.todoblock.response.CreateTodoBlockResponseDto;
+import com.knighttodo.knighttodo.rest.dto.todoblock.response.TodoBlockResponseDto;
+import com.knighttodo.knighttodo.rest.dto.todoblock.response.UpdateTodoBlockResponseDto;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
