+++ src/main/java/com/knighttodo/knighttodo/rest/mapper/TodoRestMapper.java
@@ -1,11 +1,11 @@
 package com.knighttodo.knighttodo.rest.mapper;
 
 import com.knighttodo.knighttodo.domain.TodoVO;
-import com.knighttodo.knighttodo.rest.dto.request.todo.CreateTodoRequestDto;
-import com.knighttodo.knighttodo.rest.dto.request.todo.UpdateTodoRequestDto;
-import com.knighttodo.knighttodo.rest.dto.response.todo.CreateTodoResponseDto;
-import com.knighttodo.knighttodo.rest.dto.response.todo.TodoResponseDto;
-import com.knighttodo.knighttodo.rest.dto.response.todo.UpdateTodoResponseDto;
+import com.knighttodo.knighttodo.rest.dto.todo.request.CreateTodoRequestDto;
+import com.knighttodo.knighttodo.rest.dto.todo.request.UpdateTodoRequestDto;
+import com.knighttodo.knighttodo.rest.dto.todo.response.CreateTodoResponseDto;
+import com.knighttodo.knighttodo.rest.dto.todo.response.TodoResponseDto;
+import com.knighttodo.knighttodo.rest.dto.todo.response.UpdateTodoResponseDto;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
